icon font, field previews, remove gzipped files, general refactor
