+++ README.md
@@ -20,5 +20,7 @@ jQuery(document).ready(function($) {
 ## TODO
 - Add parser to render non-editable form.
 - Add callback for `autocomplete` field.
-- Custom icons and styling
+- HTML5 fields and attributes
+- JSON data as default instead of XML
+- More examples
