+++ demo/assets/demo.css
@@ -1,6 +1,6 @@
 /*
 formBuilder - git@github.com:kevinchappell/formBuilder.git
-Version: 1.1.0
+Version: 1.2.0
 Author: Kevin Chappell <kevin.b.chappell@gmail.com>
 */
 dl dd,dl p,ol,ul{padding-left:20px}dl dd,dl p,em{font-style:italic}article,aside,details,em,figcaption,figure,footer,header,hgroup,menu,nav,pre code,section{display:block}#forkme_banner,#project_title,dl dt,h1,strong{font-weight:700}#project_tagline,#project_title{color:#fff;text-shadow:#111 0 0 10px}a,abbr,acronym,address,applet,article,aside,audio,b,big,blockquote,body,canvas,caption,center,cite,code,dd,del,details,dfn,div,dl,dt,em,embed,fieldset,figcaption,figure,footer,form,h1,h2,h3,h4,h5,h6,header,hgroup,html,i,iframe,img,ins,kbd,label,legend,li,mark,menu,nav,object,ol,output,p,pre,q,ruby,s,samp,section,small,span,strike,strong,sub,summary,sup,table,tbody,td,tfoot,th,thead,time,tr,tt,u,ul,var,video{margin:0;padding:0;border:0;font:inherit;vertical-align:baseline}blockquote,table{margin-bottom:20px}code,pre{padding:1%;width:98%;box-shadow:0 0 10px rgba(0,0,0,.1)}td,th{padding:10px}body,h6{font-size:16px}body{box-sizing:border-box;color:#373737;background:#212121;font-family:'Myriad Pro',Calibri,Helvetica,Arial,sans-serif;line-height:1.5;-webkit-font-smoothing:antialiased}h1,h2,h3,h4,h5,h6{margin:10px 0;font-weight:700;color:#222;font-family:'Lucida Grande',Calibri,Helvetica,Arial,sans-serif;letter-spacing:-1px}h1{font-size:36px}h2{padding-bottom:10px;font-size:32px;background:url(images/bg_hr.png) bottom repeat-x}h3{font-size:24px}h4{font-size:21px}h5{font-size:18px}p{margin:10px 0 15px}footer p{color:#f2f2f2}a{text-decoration:none;color:#0090ff;text-shadow:none;-webkit-transition:text-shadow .5s ease;transition:text-shadow .5s ease}a:focus,a:hover{text-decoration:underline}em{color:#999}p img{display:inline;margin:0;padding:0;vertical-align:middle;text-align:center;border:none}code,pre{color:#222;background-color:#fff;font-family:Monaco,"Bitstream Vera Sans Mono","Lucida Console",Terminal,monospace;font-size:14px;border-radius:2px}table,td{border:1px solid #373737}pre{overflow:auto;white-space:pre-wrap}code{margin:0 3px}pre code{box-shadow:none}blockquote{color:#666;padding:0 0 0 20px;border-left:3px solid #bbb}dl,ol,ul{margin-bottom:15px}ul{list-style:disc}ol{list-style:decimal}hr{height:1px;margin-bottom:5px;border:none;background:url(images/bg_hr.png) center repeat-x}table{border-collapse:collapse;border-spacing:0;text-align:left}th{font-family:'Lucida Grande','Helvetica Neue',Helvetica,Arial,sans-serif;background:#373737;color:#fff}.outer{width:100%}.inner{position:relative;max-width:1080px;padding:20px 10px;margin:0 auto}#downloads,#forkme_banner{position:absolute;z-index:10}#forkme_banner{display:block;top:0;right:10px;padding:10px 50px 10px 10px;color:#fff;background:url(images/blacktocat.png) 95% 50% no-repeat #0090ff;box-shadow:0 0 10px rgba(0,0,0,.5);border-bottom-left-radius:2px;border-bottom-right-radius:2px}#header_wrap{background:-webkit-linear-gradient(top,#373737,#212121);background:linear-gradient(to bottom,#373737,#212121)}#header_wrap .inner{padding:50px 10px 30px}#project_title{margin:0;font-size:42px}#project_tagline{font-size:24px;font-weight:300;background:0 0}#downloads{width:210px;bottom:-40px;right:0;height:70px;background:url(images/icon_download.png) 0 90% no-repeat}.tar_download_link,.zip_download_link{display:block;width:90px;height:70px;text-indent:-5000px;overflow:hidden;float:right}.zip_download_link{background:url(images/sprite_download.png) bottom left no-repeat}.tar_download_link{background:url(images/sprite_download.png) bottom right no-repeat;margin-left:10px}.zip_download_link:hover{background:url(images/sprite_download.png) top left no-repeat}.tar_download_link:hover{background:url(images/sprite_download.png) top right no-repeat}fieldset,legend{border:1px solid #999;background:#fefefe}#main_content_wrap{background:#f2f2f2;border-top:1px solid #111;border-bottom:1px solid #111}#footer_wrap{background:#212121}@media screen and (max-width:480px){body,h5{font-size:14px}.inner,code,pre{min-width:320px;max-width:480px}#downloads{display:none}#project_title{font-size:32px}h1{font-size:28px}h2{font-size:24px}h3{font-size:21px}h4{font-size:18px}h6{font-size:12px}code,pre{font-size:11px;width:98%;padding:1%}}.donate-plea{width:50%;font-size:.75em;float:right;text-align:right}.donate-plea form{float:right}.badges{margin-top:-30px}.social{margin-top:-40px}.social .chicklets{text-indent:-9999px}fieldset{padding:20px;margin:20px 0}legend{padding:5px 10px}
