--- demo/example.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Example</title>
-  <script src="js/jquery.min.js"></script>
-  <script src="js/jquery-ui.min.js"></script>
-  <script src="js/formBuilder.js"></script>
-  <link rel="stylesheet" href="css/formBuilder.css">
-  <script>
-  jQuery(document).ready(function($) {
-    'use strict';
-    $('textarea').formBuilder();
-  });
-  </script>
-</head>
-
-<body>
-  <h1>Form Builder Example</h1>
-  <form action="">
-    <textarea name="formBuilder" id="formBuilder" cols="30" rows="10"></textarea>
-  </form>
-</body>
-
-</html>
