+++ demo/index.html
@@ -8,7 +8,7 @@
   <meta name="description" content="jQuery formBuilder: A jQuery plugin for drag and drop form creation">
   <link rel="stylesheet" type="text/css" media="screen" href="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.11.4/jquery-ui.min.css">
   <link rel="stylesheet" type="text/css" media="screen" href="assets/demo.css">
-  <link rel="stylesheet" type="text/css" media="screen" href="assets/formBuilder.css">
+  <link rel="stylesheet" type="text/css" media="screen" href="assets/form-builder.css">
   <title>jQuery formBuilder | Drag &amp; Drop Form Creation</title>
 </head>
 
@@ -40,12 +40,10 @@
           <textarea name="formBuilder" id="formBuilder" cols="30" rows="10"></textarea>
         </form>
         <br style="clear:both">
-        <h2><a id="todo" class="anchor" href="#todo" aria-hidden="true"><span class="octicon octicon-link"></span></a>TODO</h2>
-        <ul>
-          <li>Add parser to render non-editable form.</li>
-          <li>Add callback for <code>autocomplete</code> field.</li>
-          <li>Custom icons and styling</li>
-        </ul>
+        <h2><a id="note" class="anchor" href="#note" aria-hidden="true"><span class="octicon octicon-link"></span></a>Thank you</h2>
+        <p>Just want to say thanks for all the positive notes and feedback, without it formBuilder would be dead in the water so keep it up.</p>
+        <p>If you found this plugin useful or even mildly interesting why not tweet about it? <a href="https://twitter.com/share" class="twitter-share-button" data-via="kevinchappell" data-hashtags="formBuilder">Tweet</a>
+<script>!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src=p+'://platform.twitter.com/widgets.js';fjs.parentNode.insertBefore(js,fjs);}}(document, 'script', 'twitter-wjs');</script></p>
       </section>
     </div>
     <!-- FOOTER  -->
@@ -69,7 +67,7 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.11.4/jquery-ui.min.js"></script>
   <script src="//cdnjs.cloudflare.com/ajax/libs/zeroclipboard/2.2.0/ZeroClipboard.js"></script>
 
-  <script src="assets/formBuilder.js"></script>
+  <script src="assets/form-builder.js"></script>
   <script>
   jQuery(document).ready(function($) {
     'use strict';
