+++ package.json
@@ -30,12 +30,12 @@
     "codeclimate-test-reporter": "^0.1.0",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^2.3.1",
+    "gulp-babel": "^5.2.1",
     "gulp-bump": "^0.3.1",
     "gulp-concat": "^2.6.0",
     "gulp-cssmin": "^0.1.7",
     "gulp-filter": "^3.0.0",
     "gulp-git": "^1.2.4",
-    "gulp-gzip": "^1.2.0",
     "gulp-header": "^1.2.2",
     "gulp-jshint": "^1.11.2",
     "gulp-sass": "^2.0.4",
