+++ src/sass/_variables.scss
@@ -1,0 +1,10 @@
+$black: #000;
+$white: #fff;
+$off-white: #fefefe;
+$grey: #666;
+$grey-light: #ccc;
+$grey-light-30: lighten($grey, 30%);
+
+$success: #08fe00;
+$error: #C10000;
+$remove: #c66865;
